feat: LottoVendingMachine class 분리
